// src/shared/api/generated/model/importClientsResponseDto.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Yoclick API
 * API multi-centro de reservas en marca blanca. Contrato canónico para la app.
 * OpenAPI spec version: 0.1.0
 */
import type { ImportClientsResponseDtoSkippedItem } from './importClientsResponseDtoSkippedItem.ts';

export interface ImportClientsResponseDto {
  /**
     * @minimum -9007199254740991
     * @maximum 9007199254740991
     */
  createdCount: number;
  /**
     * @minimum -9007199254740991
     * @maximum 9007199254740991
     */
  updatedCount: number;
  skipped: ImportClientsResponseDtoSkippedItem[];
}